AMP-30723 -implement amp TruBudget integration CLose WF Item and mark as uneditable

+++ amp/WEB-INF/src/org/dgfoundation/amp/onepager/util/ActivityUtil.java
@@ -1127,7 +1127,10 @@ public class ActivityUtil {
 //                tmpDoc.setAmpComponentFunding(null);
                 if (tmpDoc!=null){
 //                    a.getComponentFundingDocuments().remove(tmpDoc);
-                     session.delete(tmpDoc);
+                    tmpDoc.setAmpComponentFunding(null);
+                    if (tmpDoc.getId()!=null) {
+                        session.delete(tmpDoc);
+                    }
 
                 }
